requirements.txt
openai>=1.30.0,<2.0.0
reportlab>=4.0.0
playwright>=1.44.0
gspread>=6.1.0
google-auth>=2.30.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
